moved contants to a module
